More abuse of the __init__.py mechanics to remove the AssetManager class.
Now, the __init__.py provides the storage of asset paths and instances, but
the loader (and optionally logger) are still required to be provided for the
actual loading of the assets, therefore, a 'loadAssets(...)' method was added
so that an App can provide the loader/logger to the program that imports the
assets module. Note, now that this change is in place, the assets must be
loaded before the TileView makes use of them.