Added camera component (sort of a glue between views/renderer and the user/viewer).
Condensed program::main() into a subclass of App, Porcuthon, and moved a bit
of code around.

Added a Controller 'interface' to mimic MonoBehaviour, with start() and
update() methods, and adapted WorldController to make use of it. In doing so,
the 2-second randomized tiles returned.

Added mouse cursor (loaded, not used).