"Quick and dirty" camera centering.
The camera position is now considered to be the center of the viewer, instead
of at the default position of the rendering system (top-left, bottom-left, et
cetera), implemented as such:

The viewer now provides its dimensions such that, based on the camera
position (currently hard-coded to (0, 0)), will now copy a rendering into a
new render target that pads the rendering with half-viewer-dimensions around
the rendering. This new padded rendering then is copied using the renderer to
a new render target (viewable) that is finally drawn to the viewer.

Terrible description, but run it to see what happens with the hard-coded
camera position at (0, 0) to see the padding in effect.