This commit brings up the port to the middle of episode 4.
Completed allowing the player to use the middle or right mouse buttons to pan
around the rendering.

An added geometry component, vectors.Vector2, is used to manipulate the
amount to pan, overriding both add and subtraction operators (not in-place).

Since this is currently a two-dimensional game, there is no conversion from
screen coordinates to world coordinates, as the X and Y coordinate components
are exactly the same.