Get TileView with coordinates instead of indices.
The MouseController does/should not know what the tile dimensions are. This
commit introduces an 'override' of the getTileViewAt in WorldView such that
a pair of coordinates can be used to acquire a TileView, but without breaking
the current code if used elsewhere (the added parameter defaults to a value
indicating the incoming parameters are the indices of the TileView desired).